loader: Initial commit of a PVR loader
For now, we only support decompressing textures.

This uses the static PVRTexLib library that Imagination distributes. It
would be possible to use the code distributed in the SDK to not have to
rely on a binary library, but this would only work for decompression. As
I want to be able to compress texture to PVRTC as well, I need to use
the .a from Imagination.

One can imagine that we could provide a decompression only loader with
actual source code (and thus no proprietary bits), but this will be for
another time.