loader: Support incremental loading
A lot of applications are using GdkPixbufLoader so we need to support
that interface. As PVRTexLib does not really have support for
incremental loading, we just parse the header to give the width/height
to size_func and wait until the PixufLoader is closed to decompress the
texture and prepare the pixbuf.